# Conflicts:
#	gm/src/main/java/com/jjg/game/gm/controller/GMController.java
#	hall/resources/sample/GameFunction.xlsx
#	hall/resources/sample/task.xlsx
#	hall/resources/sample/warehouse.xlsx
#	hall/src/main/java/com/jjg/game/hall/minigame/game/luckytreasure/service/LuckyTreasureService.java
#	hall/src/main/java/com/jjg/game/hall/pointsaward/PointsAwardService.java
#	poker/resources/sample/GameFunction.xlsx
#	poker/resources/sample/task.xlsx
#	poker/resources/sample/warehouse.xlsx
#	slots/resources/sample/BaseRoom.xlsx
#	slots/resources/sample/GameFunction.xlsx
#	slots/resources/sample/task.xlsx
#	slots/resources/sample/warehouse.xlsx
#	table/resources/sample/GameFunction.xlsx
#	table/resources/sample/task.xlsx
#	table/resources/sample/warehouse.xlsx

+++ core/src/main/java/com/jjg/game/core/task/condition/TaskCondition10003.java
@@ -24,10 +24,10 @@ public class TaskCondition10003 extends AbstractTaskCondition<TaskConditionParam
 
     @Override
     protected boolean checkAddProgress(TaskCfg taskCfg, TaskDetail taskDetail, TaskConditionParam10003 param) {
-        List<Integer> conditionId = taskCfg.getTaskConditionId();
-        int gameId = conditionId.get(1);
-        int checkValue = conditionId.get(2);
-        int coinId = conditionId.get(4);
+        List<Long> conditionId = taskCfg.getTaskConditionId();
+        long gameId = conditionId.get(1);
+        long checkValue = conditionId.get(2);
+        long coinId = conditionId.get(4);
         //货币类型不匹配
         if (coinId != param.getCoinId()) {
             return false;
@@ -48,6 +48,6 @@ public class TaskCondition10003 extends AbstractTaskCondition<TaskConditionParam
      */
     @Override
     protected Long getCompareValue(TaskCfg taskCfg) {
-        return Long.valueOf(taskCfg.getTaskConditionId().get(3));
+        return taskCfg.getTaskConditionId().get(3);
     }
 }
